requirements.txt
PyPDF2 
sentence_transformers
faiss-cpu
numpy
openai
pytest
python-dotenv
pyfiglet
termcolor